% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rev.R
\name{autoCorrelation}
\alias{autoCorrelation}
\title{Title}
\usage{
autoCorrelation(vintages, nbreuschgodfrey = 1, nljungbox = 1)
}
\arguments{
\item{vintages}{}

\item{nbreuschgodfrey}{}

\item{nljungbox}{}

\item{reference}{}
}
\value{

}
\description{
Title
}
