% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rev.R
\name{cointegration}
\alias{cointegration}
\title{Cointegration tests (Engle-Granger)}
\usage{
cointegration(vintages, adfk = 1)
}
\arguments{
\item{vintages}{}

\item{adfk}{}
}
\value{

}
\description{
Cointegration tests (Engle-Granger)
}
