% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rev.R
\name{print.JD.Revisions.RegressionBasedAnalysis}
\alias{print.JD.Revisions.RegressionBasedAnalysis}
\title{Title}
\usage{
\method{print}{JD.Revisions.RegressionBasedAnalysis}(rev)
}
\arguments{
\item{rev}{}
}
\value{

}
\description{
Title
}
