% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rev.R
\name{revisions}
\alias{revisions}
\title{Title}
\usage{
revisions(
  periodicity,
  input,
  selection.nrevisions = 0,
  selection.firstVintage = NULL,
  selection.lastVintage = NULL
)
}
\arguments{
\item{periodicity}{}

\item{input}{}
}
\value{

}
\description{
Title
}
