% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rev.R
\name{revisionsFromCsv}
\alias{revisionsFromCsv}
\title{Title}
\usage{
revisionsFromCsv(
  file,
  periodicity,
  regDateFormat = "\%Y.\%m.\%d",
  selection.nrevisions = 0,
  selection.start = NULL,
  selection.end = NULL,
  analysis.vertical = T
)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Title
}
