% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rev.R
\name{vintageTableFromCsv}
\alias{vintageTableFromCsv}
\title{Title}
\usage{
vintageTableFromCsv(file, periodicity, regDateFormat = "\%Y.\%m.\%d")
}
\arguments{
\item{file}{}

\item{periodicity}{}

\item{regDateFormat}{}
}
\value{

}
\description{
Title
}
